Neil,
Please help me on this one. I know that we should 'use strict' but doing so is incompatible with Closure's inheritance system which we are using to subclass TreeControl.

base.js:1502
goog.base = function(me, opt_methodName, var_args) {
  var caller = arguments.callee.caller;

I had to remove use strict to get this to work. There are surely other ways to add this touch behavior and I'll pursue them if we can't figure out a good to to call the superclass from a prototype method definition.

Thanks!

R=neil.fraser@gmail.com

Review URL: https://codereview.appspot.com/11734043